+++ src/menu/menu.controller.ts
@@ -1,7 +1,8 @@
-import { Body, Controller, Delete, Get, NotFoundException, Param, Post, Put, Query } from '@nestjs/common';
+import { BadRequestException, Body, Controller, Delete, Get, NotFoundException, Param, Post, Put, Query } from '@nestjs/common';
 import { CreateMenuDto } from './dto/create-menu.dto';
 import { UpdateMenuDto } from './dto/update-menu.dto';
 import { MenuService } from './menu.service';
+import * as mongoose from 'mongoose';
 
 @Controller('menu')
 export class MenuController {
@@ -21,11 +22,14 @@ export class MenuController {
     }
 
     @Get('get-by-menu-category/:id')
-    async findByRestaurant(
-        @Param('id') id: String,
+    async findByMenuCategory(
+        @Param('id') menuCategoryId: String,
         @Query('isPopulated') isPopulated: number
     ) {
-        const menus = await this.service.findByMenuCategory(id, isPopulated);
+        if (!mongoose.isValidObjectId(menuCategoryId)) {
+            throw new BadRequestException()
+        }
+        const menus = await this.service.findByMenuCategory(menuCategoryId, isPopulated);
         if (!menus || menus.length == 0) {
             throw new NotFoundException()
         }
@@ -37,6 +41,9 @@ export class MenuController {
         @Param('id') id: String,
         @Query('isPopulated') isPopulated: number
     ) {
+        if (!mongoose.isValidObjectId(id)) {
+            throw new BadRequestException()
+        }
         const menu = await this.service.findOne(id, isPopulated);
         if (!menu) {
             throw new NotFoundException()
@@ -51,11 +58,17 @@ export class MenuController {
 
     @Put('update/:id')
     async update(@Param('id') id: String, @Body() updateMenuDto: UpdateMenuDto) {
+        if (!mongoose.isValidObjectId(id)) {
+            throw new BadRequestException()
+        }
         return await this.service.update(id, updateMenuDto);
     }
 
     @Delete('delete/:id')
     async delete(@Param('id') id: String) {
+        if (!mongoose.isValidObjectId(id)) {
+            throw new BadRequestException()
+        }
         return await this.service.delete(id);
     }
 }
